Feature implemented
Enemy spawning works.
Spawning rate is exponential. For each enemy spawned, we choose a random enemy type (entity) and random spawn location (transform), and spawn the enemy there.
The method can be scaled to include any number of spawn locations and enemy types, though if we want to spawn enemies in all zones the behavior would have to be changed. Additionally, the number of enemies to spawn is controlled by a separate function, so it can be changed at a later point.