GeneralUpdate
Stage1 done, portal needs to be added.
Stage2 started working on but having difficulty making a 3D object for maze which we might want to leave for final release but we'll see.
Stage 3 has no updates but it'll be done relatively quickly and easy by tomorrow.
Stage 4-5 are for final if we need to add them anyway.
Stage 6 is practically done, details have been reduced a lot and it's basically a copy of the original map minus anything outside of the island.