Implementation of towers, workers and resources
Implemented untested scripts for towers, workers and resource nodes. Additionally, renamed PlayerTrainer to PlayerBarracks.

Added interactions with the aforementioned types in InputHandler and the Action scripts, as well as everywhere else I remembered to.

In InputHandler and the Action scripts, added support for hotkeys (1-9) to press buttons.

All changes are untested.