Fixed enemy spawn bug and building height bug
Enemy spawn bug: navagent being enabled from the start on enemy units made moving their position buggy. The solution was to disable the navagent, and then once the object was in place, enable it.

Building height: added a height offset to the prototype building system based on the height of the mouse, by making use of how the system was built.